requirements.txt
discord.py>=2.7.1,<3
ruamel.yaml>=0.18,<0.20
schedule>=1.2,<2
dateparser>=1.2,<2
audioop-lts>=0.2; python_version >= "3.13"
